+++ CHANGELOG.md
@@ -32,6 +32,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - Use a deterministic `HashSet` in all derives, this is needed for rust analyzer
   to work correctly.
+- Use `Provider` API for backtraces in `Error` derive.
 
 ## 0.99.10 - 2020-09-11